it works, i think so

+++ GeneticAlgorithmCardsProblem/CCardBits.h
@@ -12,18 +12,19 @@ public:
 
 	static void vCrossOver(CCardBits & cFirst, CCardBits & cOther);  // implement random point and crossover
 
-	double dGetDistance();
-	void vSetDistance(double dDistance);
+	long double dGetDistance();
+	void vSetDistance(long double dDistance);
 	bool bWasChanged();
 	void vSetChanged(bool bChanged);
 	void vMutation();
-	int iTotalSum() const;
+	//void vMutation2();  // another one, just in case, random bits string xor
+	long long iTotalSum() const;
 
-	int iTotalProduct() const;
+	long long iTotalProduct() const;
 
 private:
 	int i_size;
-	double d_distance;
+	long double d_distance;
 	bool b_was_changed;
 	vector<bool> v_bits;
 }; // class CCardBits
